Add patch submitting component
